initial work on a single game dialog
